+++ src/main/java/com/missionx/questloggers/domain/partyapplicant/enums/ApplicantStatus.java
@@ -1,4 +1,4 @@
-package com.missionx.questloggers.domain.partymember.enums;
+package com.missionx.questloggers.domain.partyapplicant.enums;
 
 public enum ApplicantStatus {
     PENDING, ACCEPTED, REJECTED
